refactor: Move users related operations into one module
