+++ backend/capellacollab/sql_models/__init__.py
@@ -9,7 +9,8 @@ from importlib import metadata
 # These import statements of the models are required and should not be removed! (SQLAlchemy will not load the models otherwise)
 import capellacollab.projects.models
 from capellacollab.config import models
-from capellacollab.sql_models import notices, users
+from capellacollab.sql_models import notices
+from capellacollab.users import models
 
 log = logging.getLogger(__name__)
